Move build_new_gate to Gates
