+++ app/models/slot.rb
@@ -1,2 +1,3 @@
 class Slot < ApplicationRecord
+  validates :gates_id, presence: true
 end
